added the admin service , remains only the test files for the report endpoints

+++ gig/src/routes/delete.ts
@@ -1,5 +1,10 @@
 import express, { Request, Response } from "express";
-import { requireAuth, NotFoundError } from "@hirafee-platforme/common";
+import {
+  requireAuth,
+  NotFoundError,
+  NotAuthorizedError,
+  currentUser,
+} from "@hirafee-platforme/common";
 import { Gig } from "../models/gig";
 
 const router = express.Router();
@@ -7,11 +12,20 @@ const router = express.Router();
 router.delete(
   "/api/gigs/:id",
   requireAuth,
+  currentUser,
   async (req: Request, res: Response) => {
     const gig = await Gig.findById(req.params.id);
 
     if (!gig) {
       throw new NotFoundError();
+    }
+
+    // Check if the authenticated user is the owner of the gig or has the role of admin
+    if (
+      gig.user.toString() !== req.currentUser!.id &&
+      req.currentUser!.role !== "admin"
+    ) {
+      throw new NotAuthorizedError();
     }
 
     await Gig.findByIdAndDelete(req.params.id);
